WoWbot version one. Allow WoWbot to set class colours in the discord channel.
There are some files that are included for posterity, especially sexualrhinoceros.py

creds.py is required to run. Make sure that you install one if you end up doing this.
I realize that doing creds.py like this is a bad habit. I'm trying to get out of it.

FFS should probably actually include the requirements.txt in the next commit.